chore: preview proposal
