--- src/views/governance/views/proposal/components/ProposalDetailForm.tsx
@@ -1,45 +1,0 @@
-import { Box, Card, Text } from 'theme-ui'
-import ProposalDetail from 'views/governance/components/ProposalDetail'
-import MDEditor from '@uiw/react-md-editor'
-import { useState } from 'react'
-import { Input } from 'components'
-import Field from 'components/field'
-import { t, Trans } from '@lingui/macro'
-import { Divider } from 'theme-ui'
-
-const ProposalDetailForm = () => {
-  const [title, setTitle] = useState('')
-  const [description, setDescription] = useState<string | undefined>('')
-
-  return (
-    <Box>
-      <Card p={4} mb={4}>
-        <Text variant="sectionTitle">
-          <Trans>Proposal description</Trans>
-        </Text>
-        <Divider my={4} mx={-4} />
-        <Field label={t`Title`} mb={3}>
-          <Input
-            value={title}
-            onChange={setTitle}
-            placeholder={t`Input proposal title`}
-          />
-        </Field>
-
-        <Text variant="subtitle" ml={3} sx={{ fontSize: 1 }} mb={2}>
-          <Trans>Description</Trans>
-        </Text>
-        <MDEditor value={description} onChange={setDescription} />
-      </Card>
-
-      <ProposalDetail
-        addresses={['0x66848645Ef1C7770a00f3fA39C1C6a17781d140E']}
-        calldatas={[
-          '0xef2b9337000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000000120000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000008e870d67f660d95d5be530380d0ec0bd388289e10000000000000000000000000000000000085d4780b73119b644ae5ecd22b3760000000000000000000000004fabb145d64652a948d72533023f6e7a623c7c5300000000000000000000000067b48d65a7d3d18e56d37e0750bd2536e5bc4eb8000000000000000000000000536308b71a0aa70e14dccb4c4250055286b8483b000000000000000000000000352690cfe703f1b0388e758453865bd4d914664f00000000000000000000000055e7dd788d697b6b169b005291a1863249ad83030000000000000000000000005d3a536e4d6dbd6114cc1ead35777bab948e364300000000000000000000000039aa39c021dfbae8fac545936693ac917d5e7563000000000000000000000000f650c3d88d12db855b8bf7d11be6c55a4e07dcc9000000000000000000000000c11b1268c1a384e55c48c2391d8d480264a3a7f40000000000000000000000002260fac5e5542a773aa44fbcfedf7c193bc2c5990000000000000000000000004ddc2d193948926d02f9b1fe9e1daa0718270ed5000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000c581b735a1688071a1746c968e0798d642ede4910000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f8002000000000000000000000000000000000000000000000000011146551c1f800200000000000000000000000000000000000000000000000001116ab655e9fecf00000000000000000000000000000000000000000000000006f05b59d3b2000000000000000000000000000000000000000000000000000006f05b59d3b2000000000000000000000000000000000000000000000000000006f05b59d3b2000000000000000000000000000000000000000000000000000006f05b59d3b200000000000000000000000000000000000000000000000000000de0b6b3a7640000',
-        ]}
-      />
-    </Box>
-  )
-}
-
-export default ProposalDetailForm
