+++ src/views/governance/views/proposal/hooks/useProposal.ts
@@ -1,12 +1,14 @@
 import { useSetAtom } from 'jotai'
 import { addTransactionAtom } from 'state/atoms'
+import useProposalTx from './useProposalTx'
 
 const useProposal = () => {
+  const tx = useProposalTx()
   const addTransaction = useSetAtom(addTransactionAtom)
 
   const handlePropose = () => {}
 
-  return { fee: 0, propose: handlePropose, isValid: true }
+  return { fee: 0, propose: handlePropose, isValid: !!tx }
 }
 
 export default useProposal
