+++ src/views/governance/views/proposal/hooks/useProposalTx.ts
@@ -1,0 +1,147 @@
+import { t } from '@lingui/macro'
+import { BasketHandlerInterface } from 'abis'
+
+import { basketAtom } from 'components/rtoken-setup/atoms'
+import { BigNumber } from 'ethers'
+import { parseEther } from 'ethers/lib/utils'
+import { useAtomValue } from 'jotai'
+import { useMemo } from 'react'
+import { useFormContext } from 'react-hook-form'
+import { rTokenContractsAtom, rTokenGovernanceAtom } from 'state/atoms'
+import { TRANSACTION_STATUS } from 'utils/constants'
+import {
+  backupChangesAtom,
+  isNewBasketProposedAtom,
+  parameterContractMapAtom,
+  parametersChangesAtom,
+  revenueSplitChangesAtom,
+  roleChangesAtom,
+} from '../atoms'
+
+const useProposalTx = () => {
+  const backupChanges = useAtomValue(backupChangesAtom)
+  const revenueChanges = useAtomValue(revenueSplitChangesAtom)
+  const parameterChanges = useAtomValue(parametersChangesAtom)
+  const roleChanges = useAtomValue(roleChangesAtom)
+  const newBasket = useAtomValue(isNewBasketProposedAtom)
+  const basket = useAtomValue(basketAtom)
+  const governance = useAtomValue(rTokenGovernanceAtom)
+  const parameterMap = useAtomValue(parameterContractMapAtom)
+  const contracts = useAtomValue(rTokenContractsAtom)
+  const { getValues } = useFormContext()
+
+  return useMemo(() => {
+    const addresses: string[] = []
+    const calls: string[] = []
+    let redemptionThrottleChange = false
+    let issuanceThrottleChange = false
+    const tokenConfig = getValues()
+
+    try {
+      for (const paramChange of parameterChanges) {
+        if (
+          paramChange.field === 'issuanceThrottleAmount' ||
+          paramChange.field === 'issuanceThrottleRate'
+        ) {
+          // Skip second param if both are changed
+          if (!issuanceThrottleChange) {
+            addresses.push(parameterMap.issuanceThrottle[0].address)
+            calls.push(
+              parameterMap.issuanceThrottle[0].interface.encodeFunctionData(
+                parameterMap.issuanceThrottle[0].method,
+                [
+                  {
+                    amtRate: parseEther(tokenConfig.issuanceThrottleAmount),
+                    pctRate: parseEther(
+                      (
+                        Number(tokenConfig.issuanceThrottleRate) / 100
+                      ).toString()
+                    ),
+                  },
+                ]
+              )
+            )
+          }
+          issuanceThrottleChange = true
+        } else if (
+          paramChange.field === 'redemptionThrottleAmount' ||
+          paramChange.field === 'redemptionThrottleRate'
+        ) {
+          // Skip second param if both are changed
+          if (!redemptionThrottleChange) {
+            addresses.push(parameterMap.redemptionThrottle[0].address)
+            calls.push(
+              parameterMap.redemptionThrottle[0].interface.encodeFunctionData(
+                parameterMap.redemptionThrottle[0].method,
+                [
+                  {
+                    amtRate: parseEther(tokenConfig.redemptionThrottleAmount),
+                    pctRate: parseEther(
+                      (
+                        Number(tokenConfig.redemptionThrottleRate) / 100
+                      ).toString()
+                    ),
+                  },
+                ]
+              )
+            )
+          }
+          redemptionThrottleChange = true
+        } else {
+        }
+      }
+
+      if (newBasket) {
+        const primaryBasket: string[] = []
+        const weights: BigNumber[] = []
+
+        // TODO: Update asset registry / Assets comp/aave
+        for (const targetUnit of Object.keys(basket)) {
+          const { collaterals, distribution, scale } = basket[targetUnit]
+
+          collaterals.forEach((collateral, index) => {
+            primaryBasket.push(collateral.address)
+
+            weights.push(
+              parseEther(
+                ((Number(distribution[index]) / 100) * Number(scale)).toFixed(
+                  18
+                )
+              )
+            )
+          })
+        }
+
+        addresses.push(contracts.basketHandler)
+        calls.push(
+          BasketHandlerInterface.encodeFunctionData('setPrimeBasket', [
+            primaryBasket,
+            weights,
+          ])
+        )
+      }
+    } catch (e) {
+      console.error('Error generating proposal call', e)
+    }
+
+    return {
+      id: '',
+      description: t`New proposal`,
+      status: TRANSACTION_STATUS.PENDING,
+      value: '0',
+      call: {
+        abi: 'governance',
+        address: governance.governor,
+        method: 'propose',
+        args: [
+          addresses,
+          new Array(calls.length).fill(0),
+          calls,
+          'change basket',
+        ], // fill with empty description
+      },
+    }
+  }, [contracts])
+}
+
+export default useProposalTx
