+++ src/views/governance/views/proposal/index.tsx
@@ -1,5 +1,5 @@
 import useToggledSidebar from 'hooks/useToggledSidebar'
-import { useAtomValue, useSetAtom } from 'jotai'
+import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import { useResetAtom } from 'jotai/utils'
 import { useEffect } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
@@ -7,6 +7,7 @@ import { rTokenParamsAtom } from 'state/atoms'
 import {
   isNewBasketProposedAtom,
   isProposalEditingAtom,
+  proposalDescriptionAtom,
   proposedRolesAtom,
 } from './atoms'
 import ConfirmProposal from './components/ConfirmProposal'
@@ -15,13 +16,14 @@ import Updater from './updater'
 
 const GovernanceProposal = () => {
   const tokenParameters = useAtomValue(rTokenParamsAtom)
-  const isEditing = useAtomValue(isProposalEditingAtom)
+  const [isEditing, setEditing] = useAtom(isProposalEditingAtom)
 
   const form = useForm({
     mode: 'onChange',
     defaultValues: tokenParameters,
   })
   const setBasketProposed = useSetAtom(isNewBasketProposedAtom)
+  const setDescription = useSetAtom(proposalDescriptionAtom)
   const resetProposedRoles = useResetAtom(proposedRolesAtom)
   useToggledSidebar()
 
@@ -29,6 +31,8 @@ const GovernanceProposal = () => {
     return () => {
       setBasketProposed(false)
       resetProposedRoles()
+      setEditing(true)
+      setDescription('')
     }
   }, [])
